+++ src/entities/Badges/types.ts
@@ -71,29 +71,6 @@ export type PolygonGasData = {
   maxFeePerGas: ethers.BigNumber
   maxPriorityFeePerGas: ethers.BigNumber
 }
-type BlockNativeEstimation = {
-  confidence: number
-  price: number
-  maxPriorityFeePerGas: number
-  maxFeePerGas: number
-}
-export type BlockNativeResponse = {
-  system: string
-  network: string
-  unit: string
-  maxPrice: number
-  currentBlockNumber: number
-  msSinceLastBlock: number
-  blockPrices: [
-    {
-      blockNumber: number
-      estimatedTransactionCount: number
-      baseFeePerGas: number
-      estimatedPrices: BlockNativeEstimation[]
-    }
-  ]
-  estimatedBaseFees: never[]
-}
 
 export function isBadgeStatus(value: string | null | undefined): boolean {
   return !!value && new Set<string>(Object.values(BadgeStatus)).has(value)
